Simplification

+++ src/main/java/org/girardsimon/day01/CalibrationDocument.java
@@ -3,7 +3,7 @@ package org.girardsimon.day01;
 import java.util.List;
 
 public record CalibrationDocument(List<Integer> calibrationValues) {
-    public long sumCalibrationValues() {
+    public int sumCalibrationValues() {
         return calibrationValues.stream().mapToInt(Integer::intValue).sum();
     }
 }
